backend/requirements.txt
opencv-python
faiss-cpu
fastapi
llava @ git+https://github.com/haotian-liu/LLaVA.git
numpy
pydantic
python-dotenv
Pillow
supabase
httpx
torch
tqdm
ultralytics
uvicorn
protobuf
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
uagents